splits tracing and metrics otlp providers
