+++ pkg/telemetry/providers/otlp/metrics.go
@@ -1,0 +1,40 @@
+// Package otlp provides OpenTelemetry Protocol (OTLP) provider implementations
+package otlp
+
+import (
+	"context"
+	"fmt"
+
+	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
+	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
+)
+
+// NewMetricReader creates an OTLP metric reader for use in a unified meter provider
+func NewMetricReader(ctx context.Context, config Config) (sdkmetric.Reader, error) {
+	if config.Endpoint == "" {
+		return nil, fmt.Errorf("OTLP endpoint is required")
+	}
+
+	exporter, err := createMetricExporter(ctx, config)
+	if err != nil {
+		return nil, fmt.Errorf("failed to create OTLP metric exporter: %w", err)
+	}
+
+	return sdkmetric.NewPeriodicReader(exporter), nil
+}
+
+func createMetricExporter(ctx context.Context, config Config) (sdkmetric.Exporter, error) {
+	opts := []otlpmetrichttp.Option{
+		otlpmetrichttp.WithEndpoint(config.Endpoint),
+	}
+
+	if len(config.Headers) > 0 {
+		opts = append(opts, otlpmetrichttp.WithHeaders(config.Headers))
+	}
+
+	if config.Insecure {
+		opts = append(opts, otlpmetrichttp.WithInsecure())
+	}
+
+	return otlpmetrichttp.New(ctx, opts...)
+}
