+++ pkg/telemetry/providers/otlp/tracing.go
@@ -5,52 +5,12 @@ import (
 	"context"
 	"fmt"
 
-	"go.opentelemetry.io/otel/exporters/otlp/otlpmetric/otlpmetrichttp"
 	"go.opentelemetry.io/otel/exporters/otlp/otlptrace/otlptracehttp"
-	sdkmetric "go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
 	sdktrace "go.opentelemetry.io/otel/sdk/trace"
 	"go.opentelemetry.io/otel/trace"
 	tracenoop "go.opentelemetry.io/otel/trace/noop"
 )
-
-// Config holds OTLP-specific configuration
-type Config struct {
-	Endpoint     string
-	Headers      map[string]string
-	Insecure     bool
-	SamplingRate float64
-}
-
-// NewMetricReader creates an OTLP metric reader for use in a unified meter provider
-func NewMetricReader(ctx context.Context, config Config) (sdkmetric.Reader, error) {
-	if config.Endpoint == "" {
-		return nil, fmt.Errorf("OTLP endpoint is required")
-	}
-
-	exporter, err := createMetricExporter(ctx, config)
-	if err != nil {
-		return nil, fmt.Errorf("failed to create OTLP metric exporter: %w", err)
-	}
-
-	return sdkmetric.NewPeriodicReader(exporter), nil
-}
-
-func createMetricExporter(ctx context.Context, config Config) (sdkmetric.Exporter, error) {
-	opts := []otlpmetrichttp.Option{
-		otlpmetrichttp.WithEndpoint(config.Endpoint),
-	}
-
-	if len(config.Headers) > 0 {
-		opts = append(opts, otlpmetrichttp.WithHeaders(config.Headers))
-	}
-
-	if config.Insecure {
-		opts = append(opts, otlpmetrichttp.WithInsecure())
-	}
-
-	return otlpmetrichttp.New(ctx, opts...)
-}
 
 // NewTracerProvider creates a standalone OTLP tracer provider
 func NewTracerProvider(ctx context.Context, config Config, res *resource.Resource) (trace.TracerProvider, error) {
